% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.data.frame.light.edsurvey.data.frame.R
\name{as.data.frame}
\alias{as.data.frame}
\alias{as.data.frame.light.edsurvey.data.frame}
\title{Coerce to a Data Frame}
\usage{
\method{as.data.frame}{light.edsurvey.data.frame}(x, ...)
}
\arguments{
\item{x}{a \code{light.edsurvey.data.frame}}

\item{...}{other arguments to be passed to \code{\link[base]{as.data.frame}}}
}
\value{
a \code{data.frame}
}
\description{
Function to coerce a \code{light.edsurvey.data.frame} to a \code{data.frame}.
}
\author{
Trang Nguyen
}
