% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS.R
\name{augment}
\alias{augment}
\title{reference: Total Jackknife Method of Variance Estimation for Combined Sample}
\usage{
augment(
  data,
  subscaleWeights,
  DBA_PVs,
  PBA_PVs,
  DBAW,
  PBAW,
  PBArepWs,
  DBArepWs,
  PVs,
  StratumVar,
  PSUvar
)
}
\description{
reference: Total Jackknife Method of Variance Estimation for Combined Sample
}
