% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbind.light.edsurvey.data.frame.R
\name{cbind.light.edsurvey.data.frame}
\alias{cbind.light.edsurvey.data.frame}
\title{Combine a data frame to an \code{light.edsurvey.data.frame}.}
\usage{
\method{cbind}{light.edsurvey.data.frame}(x, ...)
}
\arguments{
\item{x}{a \code{light.edsurvey.data.frame}. The attributes of the resulting \code{light.edsurvey.data.frame} are taken from \code{x}.}

\item{...}{one or more \code{light.edsurvey.data.frame}s, or \code{data.frame}s.}
}
\value{
a \code{light.edsurvey.data.frame} with the same attributes as the first argument, \code{x}.
}
\description{
Take a vector or \code{data.frame} and combine by columns to an \code{light.edsurvey.data.frame}.
}
\seealso{
\code{\link[base]{cbind}}
}
\author{
Michael Lee and Paul Bailey
}
