% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mml.sdf.R
\name{checkIdVar}
\alias{checkIdVar}
\title{Purpose: 
Give warning if idVar is null on on light edsurvey dataframe.
Without an idVar the users won't be able to merge PVs back to a light.edsurvey.data.frame.}
\usage{
checkIdVar(data, idVar)
}
\description{
Purpose: 
Give warning if idVar is null on on light edsurvey dataframe.
Without an idVar the users won't be able to merge PVs back to a light.edsurvey.data.frame.
}
