% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mml.sdf.R
\name{checkPsuStrata}
\alias{checkPsuStrata}
\title{Give error if Strata and PSU var not present on light edsurvey data frame.}
\usage{
checkPsuStrata(data)
}
\arguments{
\item{data}{an \code{edsurvey.data.frame} or \code{light.edsurvey.data.frame} to check for PSU and Strata variables}
}
\description{
Give error if Strata and PSU var not present on light edsurvey data frame.
}
\keyword{internal}
