% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim.edsurvey.data.frame.R
\name{dim.edsurvey.data.frame}
\alias{dim.edsurvey.data.frame}
\title{Return the dimensions of an edsurvey.data.frame.}
\usage{
\method{dim}{edsurvey.data.frame}(x, ...)
}
\arguments{
\item{x}{an \code{edsurvey.data.frame}.}

\item{...}{not used, kept for compatibility}
}
\value{
returns a numeric vector of length 2, with the first element
        being the number  of rows and the second element being the number
        of columns.
}
\description{
Return the dimensions of an \code{edsurvey.data.frame}.
}
\author{
Paul Bailey
}
