% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadICCS.R
\name{downloadCivEDICCS}
\alias{downloadCivEDICCS}
\title{Instructions for Downloading and Unzipping CivED or ICCS files}
\usage{
downloadCivEDICCS(year = c(1999, 2009))
}
\arguments{
\item{year}{an integer vector indicating the study year. Valid years are 1999 and 2009.}
}
\description{
Provides instructions to download CivED or ICCS data to be processed in \code{readCivEDICCS}.
}
\seealso{
\code{\link{readCivEDICCS}}
}
\author{
Tom Fink
}
