% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadICILS.R
\name{downloadICILS}
\alias{downloadICILS}
\title{Instructions for Downloading and Unzipping ICILS Files}
\usage{
downloadICILS(years = c(2013))
}
\arguments{
\item{years}{an integer vector indicating the study year. Valid year is 2013 only.}
}
\description{
Provides instructions to download ICILS data to be processed in \code{readICILS}.
}
\examples{
\dontrun{
# view instructions to manually download study data
downloadICILS()
}
}
\seealso{
\code{\link{readICILS}}
}
\author{
Tom Fink
}
