% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadPIAAC.R
\name{downloadPIAAC}
\alias{downloadPIAAC}
\title{Download and Unzip PIAAC Files}
\usage{
downloadPIAAC(root, round = 1, cache = FALSE, verbose = TRUE)
}
\arguments{
\item{root}{a character string indicating the directory where the PIAAC data should
be stored. Note that files are placed in a folder named PIAAC and then in
a year subdirectory.}

\item{round}{a numeric value indicating the assessment round to download. Available round now is only 1 for 2012/2014 (the default).}

\item{cache}{a logical value set to \code{FALSE} to cache .txt versions of files. If set to \code{TRUE}, the function will
process all downloaded files, which might take several hours.}

\item{verbose}{a logical value that determines if you want verbose output while the function is running to indicate the progress.
Defaults to \code{TRUE}.}
}
\description{
Uses a connection to download PIAAC data to a
             computer. Data come from the OECD website.
}
\author{
Trang Nguyen
}
