% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadPISA_YAFS.R
\name{downloadPISA_YAFS}
\alias{downloadPISA_YAFS}
\title{Instructions for Downloading and Unzipping PISA YAFS Files}
\usage{
downloadPISA_YAFS(years = c(2016))
}
\arguments{
\item{years}{an integer vector indicating the study year. Valid year is 2016 only.}
}
\description{
Provides instructions to download PISA YAFS data to be processed in \code{readPISA_YAFS}.
}
\examples{
\dontrun{
# view instructions to manually download study data
downloadPISA_YAFS()
}
}
\seealso{
\code{\link{readPISA_YAFS}}
}
\author{
Tom Fink
}
