% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadTALIS.R
\name{downloadTALIS}
\alias{downloadTALIS}
\title{Instructions for Downloading TALIS Files}
\usage{
downloadTALIS(year)
}
\arguments{
\item{year}{a numeric value indicating the assessment year to download. Available years are 2008 and 2013.}
}
\description{
Provides instructions to download TALIS data to be processed in \code{\link{readTALIS}}.
}
\author{
Trang Nguyen
}
