% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mml.sdf.R
\name{es_recode}
\alias{es_recode}
\title{variable recoding function}
\usage{
es_recode(data, vars, from, to)
}
\arguments{
\item{data}{a data frame to be edited}

\item{vars}{variables to modify}

\item{from}{recode from these levels}

\item{to}{recode to this level, or levels}
}
\value{
the data with each variable in \code{vars} recoded from \code{missingFrom} to \code{missingTo}
}
\description{
Prep variables for calls to \code{\link{mml.sdf}}
}
\details{
If \code{to} is length 1, then all variables in \code{vars} are recoded from every \code{from}
to the level of \code{to}.

When \code{to} is the same length as \code{from} then the ith level of \code{from} is recoded
to the ith level of \code{to}.
}
