% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS.R
\name{getEstIV}
\alias{getEstIV}
\title{get estimate, imputation variance}
\usage{
getEstIV(data, subscaleWeights, DBA_PVs, PBA_PVs, stat, DBAW, PBAW, DBASS)
}
\arguments{
\item{data, }{the data frame with data for PBA and DBA on it}

\item{subscaleWeights}{vector of weights for subscales, in same order as DBA_PVs and PBA_PVs}

\item{DBA_PVs}{list with K elements (K=length of subscaleWeights), each element is a vector with 20 elements (one per PV)}

\item{PBA_PVs}{same as DBA_PVs, but for paper based assessments}

\item{stat}{a function that gets passed the vector of PVs and weights and retunrs the statistic}

\item{DBAW}{the full sample weight for DBA}

\item{PBAW}{the full sample weight for PBA}

\item{DBASS}{vector that subsets data to just DBA sample, rows not DBA subsample are assumed to be PBA subsample}
}
\description{
get estimate, imputation variance
}
