% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS.R
\name{getSV}
\alias{getSV}
\title{get sampling variance}
\usage{
getSV(data, subscaleWeights, DBA_PVs, PBA_PVs, stat, DBAW, PBAW, DBASS)
}
\arguments{
\item{data, }{the data frame with data for PBA and DBA on it}

\item{subscaleWeights}{vector of weights for subscales, in same order as DBA_PVs and PBA_PVs}

\item{DBA_PVs}{vector of varialble names of first DBA plausible value for each subscale}

\item{PBA_PVs}{vector of varialble names of first PBA plausible value for each subscale}

\item{stat}{a function that gets passed the vector of PVs and weights and retunrs the statistic}

\item{DBAW}{vector of 62 jackknife replicate weights for DBA data}

\item{PBAW}{vector of 62 jackknife replicate weights for PBA data}

\item{DBASS}{vector that subsets data to just DBA sample, rows not DBA subsample are assumed to be PBA subsample}
}
\description{
get sampling variance
}
