% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS.R
\name{getSVa}
\alias{getSVa}
\title{sampeling variance}
\usage{
getSVa(data, stat, rwgt, T0 = NULL, subscale = NULL, PV1 = NULL)
}
\arguments{
\item{data}{the dataset after being passed to augment}
}
\description{
sampeling variance
}
