% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.edsurvey.data.frame.R
\name{names.edsurvey.data.frame}
\alias{names.edsurvey.data.frame}
\title{Return the names of an edsurvey.data.frame.}
\usage{
\method{names}{edsurvey.data.frame}(x, ...)
}
\arguments{
\item{x}{an \code{edsurvey.data.frame}.}

\item{...}{not used. Included only for compatability.}
}
\value{
A vector showing the names of variables in an \code{edsurvey.data.frame}.
}
\description{
Return the column names of an \code{edsurvey.data.frame}.
}
\author{
Michael Lee and Paul Bailey
}
