% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm.sdf.R
\name{oddsRatio}
\alias{oddsRatio}
\title{Odds Ratios for edsurveyGlm Models}
\usage{
oddsRatio(model, alpha = 0.05)
}
\arguments{
\item{model}{an \code{edsurveyGlm} model}

\item{alpha}{the alpha level for the confidence level}
}
\value{
An \code{oddsRatio.edsurveyGlm} object with the following elements:
   \item{OR}{odds ratio coefficient estimates}
   \item{2.5\%}{lower bound 95\% confidence interval}
   \item{97.5\%}{upper bound 95\% confidence interval}
}
\description{
Converts coefficients from \code{edsurveyGlm} logit regression model to odds ratios.
}
