% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/achievementLevels.R
\name{print.achievementLevels}
\alias{print.achievementLevels}
\title{Print out a summary of the results of achievementLevels.}
\usage{
\method{print}{achievementLevels}(x, printCall = TRUE, printDiscrete = TRUE,
  printCumulative = TRUE, ...)
}
\arguments{
\item{x}{an \code{achievementLevels} object.}

\item{printCall}{logical. Set to \code{TRUE} to see details about plausible values and weights used for
calculating achievement levels.}

\item{printDiscrete}{logical. Set to \code{TRUE} to print out discrete achievement levels if they are present in \code{x}.}

\item{printCumulative}{logical. Set to \code{TRUE} to print out cumulative achievement levels if they are present in \code{x}}

\item{...}{these arguments are not passed anywhere and are included only for compatibility.}
}
\value{
A pasted statement vector that shows the details of calculated achievement levels
}
\description{
Prints details of discrete and cumulative achievement levels
 calculated using weights and variance
estimates appropriate for the \code{edsurvey.data.frame}.
}
\author{
Ahmad Emad
}
