% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.edsurvey.data.frame.R
\name{print.edsurvey.data.frame}
\alias{print.edsurvey.data.frame}
\title{Print out a summary of the edsurvey.data.frame.}
\usage{
\method{print}{edsurvey.data.frame}(x, printColnames = FALSE, ...)
}
\arguments{
\item{x}{an \code{edsurvey.data.frame}.}

\item{printColnames}{logical. Set to \code{TRUE} to see all column names.}

\item{...}{these arguments are not passed anywhere and are included only for compatibility.}
}
\value{
A pasted statement vector that shows the details of an \code{edsurvey.data.frame}
}
\description{
Prints details of an \code{edsurvey.data.frame}.
}
\author{
Michael Lee and Paul Bailey
}
