% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edsurveyTable.R
\name{print.edsurveyTable}
\alias{print.edsurveyTable}
\title{Prints summary details of a summary table from an edsurvey.data.frame.}
\usage{
\method{print}{edsurveyTable}(x, digits = getOption("digits"), ...)
}
\arguments{
\item{x}{an R object representing a summary table of an \code{edsurvey.data.frame}.}

\item{digits}{number of significant figures to print.}

\item{...}{these arguments are not passed anywhere and are included only for compatibility.}
}
\description{
Prints summary details of a summary table
estimates appropriate for the \code{edsurvey.data.frame}.
}
\author{
Paul Bailey and Howard Huo
}
