% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readHSB_SO.R
\name{readHSB_Sophomore}
\alias{readHSB_Sophomore}
\title{Connect to HS&B Study Sophomore Data}
\usage{
readHSB_Sophomore(
  HSO8092_PRI_FilePath,
  HSO8092_SASSyntax_Path,
  forceReread = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{HSO8092_PRI_FilePath}{a character value to the main study-derived
analytical data file (HSO8092_REV.PRI).
Located within the \code{REVISED_ASCII} folder.}

\item{HSO8092_SASSyntax_Path}{a character value to the SAS syntax file for
parsing the \code{HSO8092_REV.PRI} data file.
Located within the \code{SAS_EXTRACT_LOGIC} folder.}

\item{forceReread}{a logical value to force rereading of all processed data.
The default value of \code{FALSE} will speed up the read
function by using existing read-in data already processed.}

\item{verbose}{a logical value that will determine if you want verbose
output while the \code{readHSB_Sophomore} function is running
to indicate processing progress.
The default value is \code{TRUE}.}
}
\value{
an \code{edsurvey.data.frame} for the HS&B Sophomore 1980--1992 longitudinal dataset
}
\description{
Opens a connection to a High School & Beyond 1980--1992 Sophomore cohort data file and
             returns an \code{edsurvey.data.frame} with
             information about the file and data.
}
\details{
Reads in the specified \code{HSO8092_SASSyntax_Path} file to parse
the \code{HSO8092_PRI_FilePath} file.
A cached data file and metadata file will be saved in the same
directory and filename as the \code{HSO8092_PRI_FilePath} file,
having new file extensions of .txt and .meta, respectively.

Please note the original source \code{repcode} variable has been split
into two variables named \code{repcode_str} for the stratum value
and \code{repcode_psu} for the primary sampling unit (PSU) value in the resulting cache data.
}
\examples{
\dontrun{
wrkFld <- "~/HSB/SOPHOMORE"

dataPath <- file.path(wrkFld, "REVISED_ASCII", "HSO8092_REV.PRI")
sasPath <- file.path(wrkFld, "SAS_EXTRACT_LOGIC", "HSBso_READ_HSO8092.SAS")

# with verbose output as default
hsbSO <- readHSB_Sophomore(dataPath, sasPath)

# silent output
hsbSO <- readHSB_Sophomore(dataPath, sasPath, verbose = FALSE)

# force cache update
hsbSO <- readHSB_Sophomore(dataPath, sasPath, forceReread = TRUE)
}
}
\seealso{
\code{\link{readECLS_K2011}}, \code{\link{readNAEP}}, and \code{\link{getData}}
}
\author{
Tom Fink
}
