% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es_round.R
\name{roundNCES}
\alias{roundNCES}
\title{rounding helper for NCES}
\usage{
roundNCES(n)
}
\arguments{
\item{n}{the value to be rounded; accepts a vector}
}
\value{
the rounded value
}
\description{
rounding helper for NCES
}
\author{
Paul Bailey
}
