% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readNAEP.r
\name{scoreDefault}
\alias{scoreDefault}
\title{Assessment scoring}
\usage{
scoreDefault(edf, polyParamTab, dichotParamTab, scoreDict)
}
\arguments{
\item{edf}{the data}

\item{polyParamTab}{see \code{\link{mml.sdf}}}

\item{dichotParamTab}{see \code{\link{mml.sdf}}}

\item{scoreDict}{a data frame; see Details.}
}
\value{
a data frame with the columns in the \code{scoreDict} \code{key} column mapped from \code{answer} to \code{score}.
}
\description{
Score assessments
}
\details{
default scorer
scores column on edf identified by polyParamTab$ItemID, dichotParamTab$ItemID using a crosswalk in scoreDict


the \code{scoreDict} is a data frame in long format with columns \code{key}, \code{answer}, and \code{score}.
the function maps, within the item identified by \code{key} from \code{answer} to \code{score}.
}
\seealso{
\code{\link{mml.sdf}}
}
\author{
Paul Bailey and  Tom Fink
}
