% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mml.scoreTIMSS.R
\name{setTIMSSScoreDict}
\alias{setTIMSSScoreDict}
\title{setTIMSSScoreCard}
\usage{
setTIMSSScoreDict(data, subject, scoreDict = NULL)
}
\arguments{
\item{data}{a TIMSS edsurvey.data.frame}

\item{subject}{either 'mmat' for mathematics or 'ssci' for science}

\item{scoreDict}{a data frame that includes guidelines for scoring the provided TIMSS data. 
Here, \emph{scoring} refers to turning item responses into scores on each item.
To see the default scoring guidelines, call the function \code{defaultTIMSSScoreDict()}.}
}
\value{
a TIMSS edsurvey.data.frame with updated attributes
}
\description{
sets necessary attributes to run \code{mml.sdf} on TIMSS data
}
