% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.light.edsurvey.data.frame.R
\name{[.light.edsurvey.data.frame}
\alias{[.light.edsurvey.data.frame}
\title{Subset the rows of a \code{light.edsurvey.data.frame}.}
\usage{
\method{[}{light.edsurvey.data.frame}(x, i, ...)
}
\arguments{
\item{x}{a \code{light.edsurvey.data.frame}.}

\item{i}{not used. Included only for compatability.}

\item{...}{not used. Included only for compatability.}
}
\value{
An object of the same class as \code{x}, with the same attributes, that has been subsetted.
}
\description{
Return a new \code{light.edsurvey.data.frame} which, when passed to \code{getData}, returns only a subset of the data.
}
\examples{
# read in the example data (generated, not real student data)
sdf <- readNAEP(system.file("extdata/data", "M36NT2PM.dat", package = "NAEPprimer"))
gddat <- getData(sdf, c("composite", "dsex"), addAttributes = TRUE)
table(gddat[gddat$dsex=="Male","dsex"])

}
\author{
Paul Bailey
}
