% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readUTILS.R
\name{UnclassCols}
\alias{UnclassCols}
\title{remove non-standard classes from data frame columns}
\usage{
UnclassCols(tbl)
}
\description{
removes the haven defined column classes from a returned tibble
needed for issues revolving around `user_na=TRUE` argument of haven 'read_sav' method parameter
}
\keyword{internal}
