% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readNHES.R
\name{viewNHES_SurveyCodes}
\alias{viewNHES_SurveyCodes}
\title{View NHES Survey Code Definitions}
\usage{
viewNHES_SurveyCodes()
}
\description{
This function prints the defined NHES Survey Codes to console output that are compatible with the \code{readNHES} function for use.
             Typically a user will only need to manually set these codes if the 'auto' survey parameter is not able to correctly identify the
             correct survey type, or for other unusual situations.
}
\examples{
\dontrun{
  #print the NHES survey information to the console for quick reference
  viewNHES_SurveyCodes()
}
}
\seealso{
\code{readNHES}, \code{getNHES_SurveyInfo}
}
\author{
Tom Fink
}
