% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mml.sdf.R
\name{checkPsuStrata}
\alias{checkPsuStrata}
\title{Purpose: 
Give error if Strata and PSU var not present on light edsurvey dataframe.
We may want to change these to warnings later.
But leaving as errors right now. Otherwise we'll need to change getData() call to not use 
getStatumVar() and getPSUVar() functions.}
\usage{
checkPsuStrata(data)
}
\description{
Purpose: 
Give error if Strata and PSU var not present on light edsurvey dataframe.
We may want to change these to warnings later.
But leaving as errors right now. Otherwise we'll need to change getData() call to not use 
getStatumVar() and getPSUVar() functions.
}
