% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadTALIS.R
\name{downloadTALIS}
\alias{downloadTALIS}
\title{Instructions for Downloading TALIS Files}
\usage{
downloadTALIS(years)
}
\arguments{
\item{years}{a numeric value indicating the assessment year. Available years are 2008 and 2013.}
}
\description{
Provides instructions to download TALIS data to be processed in \code{\link{readTALIS}}.
}
\examples{
\dontrun{
# Print out downloading instructions for TALIS 2008 database
downloadTALIS(2008)
}
}
\seealso{
\code{\link{readTALIS}}
}
\author{
Trang Nguyen
}
