% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawPVs.sdf.R
\name{drawPVs.sdf}
\alias{drawPVs.sdf}
\title{Draw plausible values from an mml fit}
\usage{
\method{drawPVs}{sdf}(
  data,
  mml,
  npv = 5L,
  stochasticBeta = FALSE,
  normalApprox = TRUE,
  newStuDat = NULL,
  newStuItems = NULL,
  returnPosterior = FALSE,
  construct = NULL,
  ...
)
}
\arguments{
\item{data}{an \code{edsurvey.data.frame} or \code{light.edsurvey.data.frame} to which the plausible values will be added.}

\item{mml}{a fit from a call to \code{\link{mml.sdf}}, or a \code{summary.mml.sdf}, which is a  \code{summary} of 
\code{mml.sdf} call.}

\item{npv}{integer indicating the number of plausible values to draw}

\item{stochasticBeta}{logical when \code{TRUE} the regressopm coefficients will be drawn from their posterior distribution. Can also be a data frame of values (see Details).}

\item{returnPosterior}{logical set to \code{TRUE} to change output to include two additional data frames (see Value).}

\item{...}{additional parameters}
}
\description{
Draw plausible values from an mml fit
}
