% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getItems.R
\name{getAllItems}
\alias{getAllItems}
\title{Format AM dct File for Use with DirectEstimation}
\usage{
getAllItems(x, npv, construct, omittedLevels = FALSE, ...)
}
\arguments{
\item{dct}{a file location from which to read the \code{dct} file}

\item{mml}{a logical for if the paramTab is being used in \code{mml.sdf}}
}
\value{
a \code{data.frame} in a format suitable for use with \code{mml} as
a \code{paramTab}.
}
\description{
Takes an \code{AM dct} file and formats it for use with the \code{mml} method
as \code{paramTab}.
}
\author{
Sun-Joo Lee
}
