% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/achievementLevels.R
\name{print.achievementLevels}
\alias{print.achievementLevels}
\title{Print AchievementLevels Results}
\usage{
\method{print}{achievementLevels}(x, printCall = TRUE, printDiscrete = TRUE, printCumulative = TRUE, ...)
}
\arguments{
\item{x}{an \code{achievementLevels} object}

\item{printCall}{a logical value; by default (\code{TRUE}), prints details about plausible 
values and weights used for calculating achievement levels}

\item{printDiscrete}{a logical value; by default (\code{TRUE}), prints discrete achievement 
levels if they are present in \code{x}}

\item{printCumulative}{a logical value; by default (\code{TRUE}), prints cumulative achievement 
levels if they are present in \code{x}}

\item{...}{these arguments are not passed anywhere and are included only for compatibility}
}
\description{
Prints details of discrete and cumulative achievement levels
 calculated using weights and variance
estimates appropriate for the \code{edsurvey.data.frame}.
}
\author{
Huade Huo and Ahmad Emad
}
